style: multiple code restructures

+++ src/exceptions/InvalidPortException.java
@@ -1,13 +1,13 @@
 package exceptions;
 
 public class InvalidPortException extends RuntimeException {
-    private final short port;
-    public InvalidPortException(short port, String message) {
+    private final int port;
+    public InvalidPortException(int port, String message) {
         super(message);
         this.port = port;
     }
 
-    public short getPort() {
+    public int getPort() {
         return port;
     }
 }
